Updated style Customizable navbar implemented Minor optimizations and code cleanup made

+++ modules/page/admin/sections/manage.php
@@ -13,46 +13,50 @@ class module_page_admin_manage {
 			switch (self::$post['action']) {
 				case 'editPage':
 					classDB::store1('module_page_pages', array('title' => self::$post['pageTitle'], 'content' => self::$post['pageContent'], 'last_updated' => time()), '`id` = '.self::$post['id']);
-					module_page_admin::$pageContent['changesMade'] = true;
+					outputHandler::setLoadInfoVar('changesMade', true);
 					break;
 				case 'addPage':
-					$timeNow	= time();
 					$pageTitle	= self::$post['pageTitle'];
 					$pageContent	= self::$post['pageContent'];
 					classDB::insert('module_page_pages', array('title' => $pageTitle, 'content' => $pageContent, 'last_updated' => $timeNow));
 					$temp = classDB::select('module_page_pages', 'id', "`title` = '{$pageTitle}' AND `content` = '{$pageContent}' AND `last_updated` = '{$timeNow}'");
 					self::updatePageList();
-					module_page_admin::$pageContent['newPageId']	= classDB::mysqlToString($temp);
-					module_page_admin::$pageContent['pageList']	= self::$pageList;
-					module_page_admin::$pageContent['changesMade'] = true;
+					outputHandler::setLoadInfoVar('newPageId', classDB::mysqlToString($temp));
+					outputHandler::setLoadInfoVar('pageList', self::$pageList);
+					outputHandler::setLoadInfoVar('changesMade', true);
 					break;
 				default:
 					break;
 			}
-			module_page_admin::$pageContent['changesMade'] = true;
 		}
+		
 		if (self::$get['delete'] == 'true' && in_array(self::$get['edit'], array_keys(self::$pageList))) {
 			if (self::$get['edit'] != rubidium::$settings['404_page']['value']) {
-				$toDelete = self::$get['edit'];
-				classDB::delete('module_page_pages', "`id` = {$toDelete}");
-				module_page_admin::$pageContent['deletedPage'] = 'true';
-				self::updatePageList();
+				if (self::$get['edit'] != rubidium::$modules['page']['default_action_value']) {
+					$toDelete = self::$get['edit'];
+					classDB::delete('module_page_pages', "`id` = {$toDelete}");
+					outputHandler::setLoadInfoVar('deletedPage', true);
+					self::updatePageList();
+				} else {
+					outputHandler::setLoadInfoVar('cantDeleteIndex', true);
+				}
 			} else {
-				module_page_admin::$pageContent['cantDelete404'] = 'true';
+				outputHandler::setLoadInfoVar('cantDelete404', true);
 			}
 		}
+		
 		if (self::checkGetData()) {
 			//Process it
-			module_page_admin::$pageContent['changesMade'] = true;
+			outputHandler::setLoadInfoVar('changesMade', true);
 		}
 		if (self::checkUrlParams() && self::$get['delete'] != 'true') {
 			switch (self::$urlMode) {
 				case 'edit':
-					module_page_admin::$pageContent['pageEditInfo']	= self::getSinglePageInfo(rubidium::$request['GET']['edit']);
-					module_page_admin::$pageContent['subsection']	= 'edit';
+					outputHandler::setLoadInfoVar('pageEditInfo', self::getSinglePageInfo(rubidium::$request['GET']['edit']));
+					outputHandler::setLoadInfoVar('subsection', 'edit');
 					break;
 				case 'add':
-					module_page_admin::$pageContent['subsection']	= 'add';
+					outputHandler::setLoadInfoVar('subsection', 'add');
 					break;
 				default:
 					break;
@@ -67,10 +71,7 @@ class module_page_admin_manage {
 	 * Sets self::$pageList to array of all pages: id => array(data)
 	 */
 	function updatePageList() {
-		$pageList = classDB::getTable('module_page_pages', 'id', '*', '', '');
-		//print_r($pageList);
-		//print_r(array_keys($pageList));
-		self::$pageList = $pageList;
+		self::$pageList = classDB::getTable('module_page_pages', 'id', '*', '', '');
 	}
 	
 	/**
@@ -89,17 +90,33 @@ class module_page_admin_manage {
 	function checkPostData() {
 		switch (self::$post['action']) {
 			case 'editPage':
-				if (self::$post['pageTitle'] != '' && self::$post['pageContent'] != '' && self::$post['id'] != '') {
-					return true;
+				if (self::$post['pageTitle'] != '') {
+					if (self::$post['pageContent'] != '') {
+						if (self::$post['id'] != '') {
+							return true;
+						} else {
+							outputHandler::setLoadInfoVar('error', 'A valid page ID was not specified. Please check your configuration.');
+							return false;
+						}
+					} else {
+						outputHandler::setLoadInfoVar('error', 'You must enter some content.');
+						return false;
+					}
 				} else {
+					outputHandler::setLoadInfoVar('error', 'You must enter a title for the page.');
 					return false;
 				}
 				break;
-			//Need to add errors if title or content isn't filled in, rather than just returning false
 			case 'addPage':
-				if (self::$post['pageTitle'] != '' && self::$post['pageContent'] != '') {
-					return true;
+				if (self::$post['pageTitle'] != '') {
+					if (self::$post['pageContent'] != '') {
+						return true;
+					} else {
+						outputHandler::setLoadInfoVar('error', 'You must enter some content.');
+						return false;
+					}
 				} else {
+					outputHandler::setLoadInfoVar('error', 'You must enter a title for the page.');
 					return false;
 				}
 				break;
